proj_docker/train_models/requirements.txt
PyYAML
numpy
xgboost
pandas
scikit-learn
pyarrow